+++ programs/drift/src/instructions/user.rs
@@ -14,7 +14,7 @@ use crate::controller::orders::{cancel_orders, ModifyOrderId};
 use crate::controller::position::PositionDirection;
 use crate::controller::spot_balance::update_revenue_pool_balances;
 use crate::controller::spot_position::{
-    charge_withdraw_fee, update_spot_balances_and_cumulative_deposits,
+    update_spot_balances_and_cumulative_deposits,
     update_spot_balances_and_cumulative_deposits_with_limits,
 };
 use crate::error::ErrorCode;
@@ -28,6 +28,7 @@ use crate::instructions::optional_accounts::{
 use crate::instructions::SpotFulfillmentType;
 use crate::math::casting::Cast;
 use crate::math::liquidation::is_user_being_liquidated;
+use crate::math::margin::meets_initial_margin_requirement;
 use crate::math::margin::{
     calculate_max_withdrawable_amount, meets_maintenance_margin_requirement,
     meets_place_order_margin_requirement, meets_withdraw_margin_requirement,
@@ -55,8 +56,8 @@ use crate::state::high_leverage_mode_config::HighLeverageModeConfig;
 use crate::state::oracle::StrictOraclePrice;
 use crate::state::order_params::RFQMatch;
 use crate::state::order_params::{
-    ModifyOrderParams, OrderParams, PlaceAndTakeOrderSuccessCondition, PlaceOrderOptions,
-    PostOnlyParam,
+    parse_optional_params, ModifyOrderParams, OrderParams, PlaceAndTakeOrderSuccessCondition,
+    PlaceOrderOptions, PostOnlyParam,
 };
 use crate::state::paused_operations::{PerpOperation, SpotOperation};
 use crate::state::perp_market::ContractType;
@@ -72,7 +73,6 @@ use crate::state::spot_market_map::{
 use crate::state::state::State;
 use crate::state::swift_user::SwiftOrderId;
 use crate::state::swift_user::SwiftUserOrdersLoader;
-use crate::state::swift_user::SwiftUserOrdersZeroCopy;
 use crate::state::swift_user::{SwiftUserOrders, SWIFT_PDA_SEED};
 use crate::state::traits::Size;
 use crate::state::user::ReferrerStatus;
@@ -359,6 +359,14 @@ pub fn handle_deposit<'c: 'info, 'info>(
     let oracle_price_data = &oracle_map.get_price_data(&spot_market.oracle)?.clone();
 
     validate!(
+        user.pool_id == spot_market.pool_id,
+        ErrorCode::InvalidPoolId,
+        "user pool id ({}) != market pool id ({})",
+        user.pool_id,
+        spot_market.pool_id
+    )?;
+
+    validate!(
         !matches!(spot_market.status, MarketStatus::Initialized),
         ErrorCode::MarketBeingInitialized,
         "Market is being initialized"
@@ -569,12 +577,6 @@ pub fn handle_withdraw<'c: 'info, 'info>(
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
         let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle)?;
 
-        if user.qualifies_for_withdraw_fee(&user_stats, slot) {
-            let fee =
-                charge_withdraw_fee(spot_market, oracle_price_data.price, user, &mut user_stats)?;
-            amount = amount.safe_sub(fee.cast()?)?;
-        }
-
         user.increment_total_withdraws(
             amount,
             oracle_price_data.price,
@@ -737,6 +739,14 @@ pub fn handle_transfer_deposit<'c: 'info, 'info>(
     {
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
 
+        validate!(
+            from_user.pool_id == spot_market.pool_id,
+            ErrorCode::InvalidPoolId,
+            "user pool id ({}) != market pool id ({})",
+            from_user.pool_id,
+            spot_market.pool_id
+        )?;
+
         from_user.increment_total_withdraws(
             amount,
             oracle_price,
@@ -803,6 +813,14 @@ pub fn handle_transfer_deposit<'c: 'info, 'info>(
 
     {
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
+
+        validate!(
+            to_user.pool_id == spot_market.pool_id,
+            ErrorCode::InvalidPoolId,
+            "user pool id ({}) != market pool id ({})",
+            to_user.pool_id,
+            spot_market.pool_id
+        )?;
 
         to_user.increment_total_deposits(
             amount,
@@ -1300,7 +1318,7 @@ pub fn handle_place_orders<'c: 'info, 'info>(
 pub fn handle_place_and_take_perp_order<'c: 'info, 'info>(
     ctx: Context<'_, '_, 'c, 'info, PlaceAndTake<'info>>,
     params: OrderParams,
-    success_condition: Option<u32>, // u32 for backwards compatibility
+    optional_params: Option<u32>, // u32 for backwards compatibility
 ) -> Result<()> {
     let clock = Clock::get()?;
     let state = &ctx.accounts.state;
@@ -1340,6 +1358,8 @@ pub fn handle_place_and_take_perp_order<'c: 'info, 'info>(
     let mut user = load_mut!(ctx.accounts.user)?;
     let clock = Clock::get()?;
 
+    let (success_condition, auction_duration_percentage) = parse_optional_params(optional_params);
+
     controller::orders::place_perp_order(
         &ctx.accounts.state,
         &mut user,
@@ -1371,7 +1391,10 @@ pub fn handle_place_and_take_perp_order<'c: 'info, 'info>(
         &makers_and_referrer_stats,
         None,
         &Clock::get()?,
-        FillMode::PlaceAndTake(is_immediate_or_cancel || success_condition.is_some()),
+        FillMode::PlaceAndTake(
+            is_immediate_or_cancel || optional_params.is_some(),
+            auction_duration_percentage,
+        ),
     )?;
 
     let order_exists = load!(ctx.accounts.user)?
@@ -1390,20 +1413,18 @@ pub fn handle_place_and_take_perp_order<'c: 'info, 'info>(
         )?;
     }
 
-    if let Some(success_condition) = success_condition {
-        if success_condition == PlaceAndTakeOrderSuccessCondition::PartialFill as u32 {
-            validate!(
-                base_asset_amount_filled > 0,
-                ErrorCode::PlaceAndTakeOrderSuccessConditionFailed,
-                "no partial fill"
-            )?;
-        } else if success_condition == PlaceAndTakeOrderSuccessCondition::FullFill as u32 {
-            validate!(
-                base_asset_amount_filled > 0 && !order_exists,
-                ErrorCode::PlaceAndTakeOrderSuccessConditionFailed,
-                "no full fill"
-            )?;
-        }
+    if success_condition == PlaceAndTakeOrderSuccessCondition::PartialFill as u8 {
+        validate!(
+            base_asset_amount_filled > 0,
+            ErrorCode::PlaceAndTakeOrderSuccessConditionFailed,
+            "no partial fill"
+        )?;
+    } else if success_condition == PlaceAndTakeOrderSuccessCondition::FullFill as u8 {
+        validate!(
+            base_asset_amount_filled > 0 && !order_exists,
+            ErrorCode::PlaceAndTakeOrderSuccessConditionFailed,
+            "no full fill"
+        )?;
     }
 
     Ok(())
@@ -2191,6 +2212,39 @@ pub fn handle_update_user_margin_trading_enabled<'c: 'info, 'info>(
 
     validate_spot_margin_trading(&user, &perp_market_map, &spot_market_map, &mut oracle_map)
         .map_err(|_| ErrorCode::MarginOrdersOpen)?;
+
+    Ok(())
+}
+
+pub fn handle_update_user_pool_id<'c: 'info, 'info>(
+    ctx: Context<'_, '_, 'c, 'info, UpdateUser<'info>>,
+    _sub_account_id: u16,
+    pool_id: u8,
+) -> Result<()> {
+    #[cfg(all(feature = "mainnet-beta", not(feature = "anchor-test")))]
+    {
+        panic!("pools disabled on mainnet-beta");
+    }
+
+    let remaining_accounts_iter = &mut ctx.remaining_accounts.iter().peekable();
+    let AccountMaps {
+        perp_market_map,
+        spot_market_map,
+        mut oracle_map,
+        ..
+    } = load_maps(
+        remaining_accounts_iter,
+        &MarketSet::new(),
+        &MarketSet::new(),
+        Clock::get()?.slot,
+        None,
+    )?;
+
+    let mut user = load_mut!(ctx.accounts.user)?;
+    user.pool_id = pool_id;
+
+    // will throw if user has deposits/positions in other pools
+    meets_initial_margin_requirement(&user, &perp_market_map, &spot_market_map, &mut oracle_map)?;
 
     Ok(())
 }