+++ programs/drift/src/lib.rs
@@ -319,6 +319,14 @@ pub mod drift {
         handle_update_user_margin_trading_enabled(ctx, _sub_account_id, margin_trading_enabled)
     }
 
+    pub fn update_user_pool_id<'c: 'info, 'info>(
+        ctx: Context<'_, '_, 'c, 'info, UpdateUser<'info>>,
+        _sub_account_id: u16,
+        pool_id: u8,
+    ) -> Result<()> {
+        handle_update_user_pool_id(ctx, _sub_account_id, pool_id)
+    }
+
     pub fn update_user_delegate(
         ctx: Context<UpdateUser>,
         _sub_account_id: u16,
@@ -347,6 +355,12 @@ pub mod drift {
         ctx: Context<'_, '_, 'c, 'info, DeleteUser>,
     ) -> Result<()> {
         handle_delete_user(ctx)
+    }
+
+    pub fn force_delete_user<'c: 'info, 'info>(
+        ctx: Context<'_, '_, 'c, 'info, ForceDeleteUser<'info>>,
+    ) -> Result<()> {
+        handle_force_delete_user(ctx)
     }
 
     pub fn delete_swift_user_orders<'c: 'info, 'info>(
@@ -621,6 +635,13 @@ pub mod drift {
         ctx: Context<UpdateUserGovTokenInsuranceStake>,
     ) -> Result<()> {
         handle_update_user_gov_token_insurance_stake(ctx)
+    }
+
+    pub fn update_user_gov_token_insurance_stake_devnet(
+        ctx: Context<UpdateUserGovTokenInsuranceStakeDevnet>,
+        gov_stake_amount: u64,
+    ) -> Result<()> {
+        handle_update_user_gov_token_insurance_stake_devnet(ctx, gov_stake_amount)
     }
 
     // IF stakers
@@ -996,6 +1017,13 @@ pub mod drift {
         insurance_fund_unstaking_period: i64,
     ) -> Result<()> {
         handle_update_insurance_fund_unstaking_period(ctx, insurance_fund_unstaking_period)
+    }
+
+    pub fn update_spot_market_pool_id(
+        ctx: Context<AdminUpdateSpotMarket>,
+        pool_id: u8,
+    ) -> Result<()> {
+        handle_update_spot_market_pool_id(ctx, pool_id)
     }
 
     pub fn update_spot_market_liquidation_fee(