+++ programs/drift/src/state/perp_market.rs
@@ -41,8 +41,6 @@ use crate::state::paused_operations::PerpOperation;
 use drift_macros::assert_no_slop;
 use static_assertions::const_assert_eq;
 
-use super::oracle::PrelaunchOracle;
-
 #[cfg(test)]
 mod tests;
 
@@ -243,7 +241,7 @@ pub struct PerpMarket {
     /// fuel multiplier for perp maker
     /// precision: 10
     pub fuel_boost_maker: u8,
-    pub padding1: u8,
+    pub pool_id: u8,
     pub high_leverage_margin_ratio_initial: u16,
     pub high_leverage_margin_ratio_maintenance: u16,
     pub padding: [u8; 38],
@@ -283,7 +281,7 @@ impl Default for PerpMarket {
             fuel_boost_position: 0,
             fuel_boost_taker: 0,
             fuel_boost_maker: 0,
-            padding1: 0,
+            pool_id: 0,
             high_leverage_margin_ratio_initial: 0,
             high_leverage_margin_ratio_maintenance: 0,
             padding: [0; 38],
@@ -327,7 +325,7 @@ impl PerpMarket {
         }
 
         let amm_low_inventory_and_profitable =
-            (self.amm.net_revenue_since_last_funding > 0 && amm_lp_allowed_to_jit_make);
+            self.amm.net_revenue_since_last_funding > 0 && amm_lp_allowed_to_jit_make;
         let amm_oracle_no_latency = self.amm.oracle_source == OracleSource::Prelaunch;
         let can_skip = amm_low_inventory_and_profitable || amm_oracle_no_latency;
 
@@ -1400,10 +1398,17 @@ impl AMM {
 
     pub fn can_lower_k(&self) -> DriftResult<bool> {
         let (max_bids, max_asks) = amm::calculate_market_open_bids_asks(self)?;
-        let can_lower = self.base_asset_amount_with_amm.unsigned_abs()
-            < max_bids.unsigned_abs().min(max_asks.unsigned_abs())
-            && self.base_asset_amount_with_amm.unsigned_abs()
-                < self.sqrt_k.safe_sub(self.user_lp_shares)?;
+        let min_order_size_u128 = self.min_order_size.cast::<u128>()?;
+
+        let can_lower = (self.base_asset_amount_with_amm.unsigned_abs()
+            < max_bids.unsigned_abs().min(max_asks.unsigned_abs()))
+            && (self
+                .base_asset_amount_with_amm
+                .unsigned_abs()
+                .max(min_order_size_u128)
+                < self.sqrt_k.safe_sub(self.user_lp_shares)?)
+            && (min_order_size_u128 < max_bids.unsigned_abs().max(max_asks.unsigned_abs()));
+
         Ok(can_lower)
     }