+++ programs/drift/src/state/user.rs
@@ -5,8 +5,7 @@ use crate::math::auction::{calculate_auction_price, is_auction_complete};
 use crate::math::casting::Cast;
 use crate::math::constants::{
     EPOCH_DURATION, FUEL_START_TS, OPEN_ORDER_MARGIN_REQUIREMENT,
-    PRICE_TIMES_AMM_TO_QUOTE_PRECISION_RATIO, QUOTE_PRECISION, QUOTE_SPOT_MARKET_INDEX,
-    TEN_THOUSAND_QUOTE, THIRTY_DAY, TWENTY_FOUR_HOUR, TWO_HUNDRED_FIFTY_THOUSAND_QUOTE,
+    PRICE_TIMES_AMM_TO_QUOTE_PRECISION_RATIO, QUOTE_PRECISION, QUOTE_SPOT_MARKET_INDEX, THIRTY_DAY,
 };
 use crate::math::lp::{calculate_lp_open_bids_asks, calculate_settle_lp_metrics};
 use crate::math::margin::MarginRequirementType;
@@ -128,7 +127,8 @@ pub struct User {
     /// Whether or not user has open order with auction
     pub has_open_auction: bool,
     pub margin_mode: MarginMode,
-    pub padding1: [u8; 4],
+    pub pool_id: u8,
+    pub padding1: [u8; 3],
     pub last_fuel_bonus_update_ts: u32,
     pub padding: [u8; 12],
 }
@@ -391,19 +391,6 @@ impl User {
         }
     }
 
-    pub fn qualifies_for_withdraw_fee(&self, user_stats: &UserStats, slot: u64) -> bool {
-        // only qualifies for user with recent last_active_slot (~25 seconds)
-        if slot.saturating_sub(self.last_active_slot) >= 50 {
-            return false;
-        }
-
-        let min_total_withdraws = 10_000_000 * QUOTE_PRECISION_U64; // $10M
-
-        // if total withdraws are greater than $10M and user has paid more than %.01 of it in fees
-        self.total_withdraws >= min_total_withdraws
-            && self.total_withdraws / user_stats.fees.total_fee_paid.max(1) > 10_000
-    }
-
     pub fn update_reduce_only_status(&mut self, reduce_only: bool) -> DriftResult {
         if reduce_only {
             self.add_user_status(UserStatus::ReduceOnly);
@@ -506,6 +493,7 @@ impl User {
         let strict = margin_requirement_type == MarginRequirementType::Initial;
         let context = MarginContext::standard(margin_requirement_type)
             .strict(strict)
+            .ignore_invalid_deposit_oracles(true)
             .fuel_spot_delta(withdraw_market_index, withdraw_amount.cast::<i128>()?)
             .fuel_numerator(self, now);
 
@@ -519,7 +507,7 @@ impl User {
 
         if calculation.margin_requirement > 0 || calculation.get_num_of_liabilities()? > 0 {
             validate!(
-                calculation.all_oracles_valid,
+                calculation.all_liability_oracles_valid,
                 ErrorCode::InvalidOracle,
                 "User attempting to withdraw with outstanding liabilities when an oracle is invalid"
             )?;